Add CommandParameter
